[DOP-21576] Move /auth/me -> /users/me; Remove state from redirect
