+++ .env.local
@@ -20,7 +20,7 @@ export DATA_RENTGEN__AUTH__KEYCLOAK__SERVER_URL=http://localhost:8080
 export DATA_RENTGEN__AUTH__KEYCLOAK__REALM_NAME=manually_created
 export DATA_RENTGEN__AUTH__KEYCLOAK__CLIENT_ID=manually_created
 export DATA_RENTGEN__AUTH__KEYCLOAK__CLIENT_SECRET=change_me
-export DATA_RENTGEN__AUTH__KEYCLOAK__REDIRECT_URI=http://localhost:3000/callback
+export DATA_RENTGEN__AUTH__KEYCLOAK__REDIRECT_URI=http://localhost:3000/auth/callback
 export DATA_RENTGEN__AUTH__KEYCLOAK__SCOPE=email
 export DATA_RENTGEN__AUTH__KEYCLOAK__VERIFY_SSL=False
 export DATA_RENTGEN__AUTH__PROVIDER=data_rentgen.server.providers.auth.keycloak_provider.KeycloakAuthProvider
