+++ data_rentgen/server/api/v1/router/__init__.py
@@ -8,6 +8,7 @@ from data_rentgen.server.api.v1.router.job import router as job_router
 from data_rentgen.server.api.v1.router.location import router as location_router
 from data_rentgen.server.api.v1.router.operation import router as operation_router
 from data_rentgen.server.api.v1.router.run import router as run_router
+from data_rentgen.server.api.v1.router.user import router as user_router
 
 router = APIRouter(prefix="/v1")
 router.include_router(auth_router)
@@ -16,3 +17,4 @@ router.include_router(job_router)
 router.include_router(location_router)
 router.include_router(operation_router)
 router.include_router(run_router)
+router.include_router(user_router)
