+++ data_rentgen/server/providers/auth/keycloak_provider.py
@@ -1,13 +1,10 @@
 # SPDX-FileCopyrightText: 2024 MTS PJSC
 # SPDX-License-Identifier: Apache-2.0
-import base64
 import logging
-from http import HTTPStatus
 from typing import Annotated, Any
 
-from fastapi import Depends, FastAPI, HTTPException, Request
+from fastapi import Depends, FastAPI, Request
 from keycloak import KeycloakOpenID
-from keycloak.exceptions import KeycloakConnectionError
 
 from data_rentgen.db.models import User
 from data_rentgen.dependencies import Stub
@@ -81,7 +78,7 @@ class KeycloakAuthProvider(AuthProvider):
 
         if not access_token:
             logger.debug("No access token found in session.")
-            self.redirect_to_auth(str(request.url))
+            self.redirect_to_auth()
 
         # if user is disabled or blocked in Keycloak after the token is issued, he will
         # remain authorized until the token expires (not more than 15 minutes in MTS SSO)
@@ -96,7 +93,7 @@ class KeycloakAuthProvider(AuthProvider):
 
             if token_info is None:
                 # If there is no token_info after refresh user get redirect
-                self.redirect_to_auth(str(request.url))
+                self.redirect_to_auth()
 
         # these names are hardcoded in keycloak:
         # https://github.com/keycloak/keycloak/blob/3ca3a4ad349b4d457f6829eaf2ae05f1e01408be/core/src/main/java/org/keycloak/representations/IDToken.java
@@ -107,11 +104,7 @@ class KeycloakAuthProvider(AuthProvider):
         return await self._uow.user.get_or_create(UserDTO(name=login))  # type: ignore[arg-type]
 
     async def logout(self, refresh_token: str):
-        try:
-            return self.keycloak_openid.logout(refresh_token)
-        except KeycloakConnectionError as err:
-            logger.error("Error when trying to get token: %s", err)
-            return None
+        return self.keycloak_openid.logout(refresh_token)
 
     def decode_token(self, access_token: str) -> dict[str, Any] | None:
         try:
@@ -127,21 +120,14 @@ class KeycloakAuthProvider(AuthProvider):
             return new_tokens.get("access_token"), new_tokens.get("refresh_token")
         except Exception as err:
             logger.debug("Failed to refresh access token: %s", err)
-            self.redirect_to_auth(origin_url)
+            self.redirect_to_auth()
 
-    def redirect_to_auth(self, state: str = ""):
-        try:
-            state = base64.b64encode(state.encode("utf-8"))  # type: ignore[assignment]
+    def redirect_to_auth(self):
 
-            auth_url = self.keycloak_openid.auth_url(
-                redirect_uri=self.settings.keycloak.redirect_uri,
-                scope=self.settings.keycloak.scope,
-                state=state.decode("utf-8"),  # type: ignore[attr-defined]
-            )
+        auth_url = self.keycloak_openid.auth_url(
+            redirect_uri=self.settings.keycloak.redirect_uri,
+            scope=self.settings.keycloak.scope,
+        )
 
-        except KeycloakConnectionError as err:
-            logger.error("Failed connect to Keycloak: %s", err)
-            # TODO: What exception should we raise here?
-            raise HTTPException(status_code=HTTPStatus.BAD_REQUEST, detail=err)
         logger.info("Raising redirect error with url: %s", auth_url)
-        raise RedirectError(message="Authorize on provided url", details=auth_url)
+        raise RedirectError(message="Please authorize using provided URL", details=auth_url)
