+++ tests/test_server/test_auth/test_keycloak_auth.py
@@ -1,4 +1,3 @@
-import base64
 import logging
 
 import pytest
@@ -9,7 +8,6 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from data_rentgen.db.models import Dataset, User
 from data_rentgen.server.settings import ServerApplicationSettings as Settings
 from data_rentgen.server.settings.auth.keycloak import KeycloakSettings
-from tests.test_server.utils.enrich import enrich_datasets
 
 KEYCLOAK_PROVIDER = "data_rentgen.server.providers.auth.keycloak_provider.KeycloakAuthProvider"
 pytestmark = [pytest.mark.asyncio, pytest.mark.server]
@@ -33,23 +31,21 @@ async def test_get_keycloak_user_unauthorized(
     caplog,
     server_app_settings: Settings,
 ):
-    k_settings = KeycloakSettings.model_validate(server_app_settings.auth.keycloak)
+    settings = KeycloakSettings.model_validate(server_app_settings.auth.keycloak)
 
-    response = await test_client.get("/v1/datasets")
+    response = await test_client.get("/v1/users/me")
 
     # redirect unauthorized user to Keycloak
-    state = str(response.url).encode("utf-8")
-    state = base64.b64encode(state)
     redirect_url = (
-        f"{k_settings.server_url}/realms/{k_settings.realm_name}/protocol/openid-connect/auth?client_id="
-        f"{k_settings.client_id}&response_type=code&redirect_uri={k_settings.redirect_uri}"
-        f"&scope={k_settings.scope}&state={state.decode('utf-8')}&nonce="
+        f"{settings.server_url}/realms/{settings.realm_name}/protocol/openid-connect/auth?client_id="
+        f"{settings.client_id}&response_type=code&redirect_uri={settings.redirect_uri}"
+        f"&scope={settings.scope}&state=&nonce="
     )
     assert response.status_code == 401
     assert response.json() == {
         "error": {
             "code": "auth_redirect",
-            "message": "Authorize on provided url",
+            "message": "Please authorize using provided URL",
             "details": redirect_url,
         },
     }
@@ -71,53 +67,24 @@ async def test_get_keycloak_user_authorized(
     test_client: AsyncClient,
     async_session: AsyncSession,
     user: User,
-    datasets: list[Dataset],
     server_app_settings: Settings,
     create_session_cookie,
     mock_keycloak_well_known,
     mock_keycloak_realm,
 ):
-    datasets = await enrich_datasets(datasets, async_session)
     session_cookie = create_session_cookie(user)
     headers = {
         "Cookie": f"session={session_cookie}",
     }
 
     response = await test_client.get(
-        "/v1/datasets",
+        "/v1/users/me",
         headers=headers,
     )
 
     assert response.cookies.get("session") == session_cookie
     assert response.status_code == 200
-    assert response.json() == {
-        "meta": {
-            "page": 1,
-            "page_size": 20,
-            "total_count": len(datasets),
-            "pages_count": 1,
-            "has_next": False,
-            "has_previous": False,
-            "next_page": None,
-            "previous_page": None,
-        },
-        "items": [
-            {
-                "kind": "DATASET",
-                "id": dataset.id,
-                "format": dataset.format,
-                "name": dataset.name,
-                "location": {
-                    "id": dataset.location.id,
-                    "name": dataset.location.name,
-                    "type": dataset.location.type,
-                    "addresses": [{"url": address.url} for address in dataset.location.addresses],
-                    "external_id": dataset.location.external_id,
-                },
-            }
-            for dataset in sorted(datasets, key=lambda x: x.name)
-        ],
-    }
+    assert response.json() == {"name": user.name}
 
 
 @responses.activate
@@ -135,7 +102,6 @@ async def test_get_keycloak_user_authorized(
 async def test_get_keycloak_user_expired_access_token(
     caplog,
     user: User,
-    datasets: list[Dataset],
     test_client: AsyncClient,
     async_session: AsyncSession,
     server_app_settings: Settings,
@@ -144,7 +110,6 @@ async def test_get_keycloak_user_expired_access_token(
     mock_keycloak_realm,
     mock_keycloak_token_refresh,
 ):
-    datasets = await enrich_datasets(datasets, async_session)
     session_cookie = create_session_cookie(user, expire_in_msec=-100000000)  # expired access token
     headers = {
         "Cookie": f"session={session_cookie}",
@@ -152,37 +117,10 @@ async def test_get_keycloak_user_expired_access_token(
 
     with caplog.at_level(logging.DEBUG):
         response = await test_client.get(
-            "/v1/datasets",
+            "/v1/users/me",
             headers=headers,
         )
 
     assert response.cookies.get("session") != session_cookie, caplog.text  # cookie is updated
     assert response.status_code == 200
-    assert response.json() == {
-        "meta": {
-            "page": 1,
-            "page_size": 20,
-            "total_count": len(datasets),
-            "pages_count": 1,
-            "has_next": False,
-            "has_previous": False,
-            "next_page": None,
-            "previous_page": None,
-        },
-        "items": [
-            {
-                "kind": "DATASET",
-                "id": dataset.id,
-                "format": dataset.format,
-                "name": dataset.name,
-                "location": {
-                    "id": dataset.location.id,
-                    "name": dataset.location.name,
-                    "type": dataset.location.type,
-                    "addresses": [{"url": address.url} for address in dataset.location.addresses],
-                    "external_id": dataset.location.external_id,
-                },
-            }
-            for dataset in sorted(datasets, key=lambda x: x.name)
-        ],
-    }
+    assert response.json() == {"name": user.name}
